Added contactMaterial to the Ned robot's jaw and contactProperty to the worldInfo
To increase the friction of the robot's jaw we need a new contactMaterial to define the friction between this new material and everything else. Also, it is necessary to have our own PROTO file of the Ned robot to make the desired changes. Thus first commit of an own PROTO file including the new jaw_material property of the jaw.

In the wbt file you can find the definition of a new contactProperty defining the coulombFriction between the jaw_material and everything else (no material2 is defined).